Controller: fix circular dependency
